完成消息类型策略类

+++ talktime-framework/talktime-dao/src/main/java/com/qingmeng/entity/ChatMessage.java
@@ -61,7 +61,7 @@ public class ChatMessage implements Serializable {
     /**
      * 与回复消息的间隔条数
      */
-    private Integer gapCount;
+    private Long gapCount;
 
     /**
      * 消息类型
